requirements.txt
pandas
fsspec
google-cloud-storage
datetime
requests
gcsfs
